editor/lsp: optimized the transaction sent for didChange.
Add new change triggerer `Undo` and `Paste` to differentiate these two.

It let the editor methods know if they have to send a didChange notification
to the LSP server: if it's not .Input, they don't, the Undo, Redo and Paste
method are responsible of sending the didChange notification once they're done.